Create adsr-class (1).zip
This is an adsr class zip file containing multiple cpp and h files. Some such as filter, debouncer, wavetable, may be replaced. This adsr currently uses GUI controls to modify adsr paramaters, which will be replaced with analog ins. It uses digital 0 to trigger the adsr. Additionally, it contains a filter ADSR, so two adsrs. one affecting the level, and the other controlling the filter.